massive speed increase on get methods.
added a method to the parser that only parses a single header and the jstore class uses that for most of its methods that dont write to the jstore file like get(), keys(), headerExists() and so on. before a get method on a jstore file with 50k lines would take 110ms now it only takes 9 ms.